third_party/WebKit/LayoutTests/fast/js/toString-overrides-expected.txt

This test checks for regression against 4147: Array.toString() and toLocaleString() improvements from KDE KJS.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS [1].toString() is '1'
PASS [1].toLocaleString() is 'toLocaleString'
FAIL [1].toLocaleString() should be 1. Threw exception TypeError: Property 'toLocaleString' of object [object Number] is not a function
PASS [/r/].toString() is 'toString2'
PASS [/r/].toLocaleString() is 'toLocaleString2'
FAIL [/r/].toLocaleString() should be toString2. Threw exception TypeError: Property 'toLocaleString' of object [object RegExp] is not a function
PASS caught is true
PASS successfullyParsed is true

TEST COMPLETE
